+++ android/src/main/java/com/relateddigital/flutter/Constants.java
@@ -30,6 +30,7 @@ public class Constants {
     public static String M_GET_EXVISITORID = "VL/getExVisitorID";
     public static String M_APP_TRACKER = "VL/appTracker";
     public static String M_GET_PUSH_MESSAGES = "RD/getPushMessages";
+    public static String M_SEND_LOCATION_PERMISSION = "VL/sendLocationPermission";
 
     public static String VL_CHANNEL = "Android";
     public static String VL_SEGMENT_URL = "https://lgr.visilabs.net";