+++ example/ios/NotificationContent/EMNotificationViewController.swift
@@ -1,0 +1,65 @@
+//
+//  EMNotificationViewController.swift
+//  NotificationContent
+//
+//  Created by Egemen Gülkılık on 3.12.2021.
+//
+
+import UIKit
+import UserNotifications
+import UserNotificationsUI
+import Euromsg
+
+@objc(EMNotificationViewController)
+class EMNotificationViewController: UIViewController, UNNotificationContentExtension {
+    
+    let carouselView = EMNotificationCarousel.initView()
+    var completion: ((_ url: URL?, _ bestAttemptContent: UNMutableNotificationContent?) -> Void)?
+    
+    var notificationRequestIdentifier = ""
+    
+    func didReceive(_ notification: UNNotification) {
+        notificationRequestIdentifier = notification.request.identifier
+        Euromsg.configure(appAlias: "relateddigital-flutter-example-ios", launchOptions: nil, enableLog: true)
+        carouselView.didReceive(notification)
+    }
+    func didReceive(_ response: UNNotificationResponse, completionHandler completion: @escaping (UNNotificationContentExtensionResponseOption) -> Void) {
+        carouselView.didReceive(response, completionHandler: completion)
+
+    }
+    override func loadView() {
+        completion = { [weak self] url, bestAttemptContent in
+            if let identifier = self?.notificationRequestIdentifier {
+                UNUserNotificationCenter.current().removeDeliveredNotifications(withIdentifiers: [identifier])
+                UNUserNotificationCenter.current().getDeliveredNotifications(completionHandler: { notifications in
+                    bestAttemptContent?.badge =  NSNumber(value: notifications.count)
+                })
+            }
+            if let url = url {
+                if #available(iOSApplicationExtension 12.0, *) {
+                    self?.extensionContext?.dismissNotificationContentExtension()
+                }
+                self?.extensionContext?.open(url)
+            } else {
+                if #available(iOSApplicationExtension 12.0, *) {
+                    self?.extensionContext?.performNotificationDefaultAction()
+                }
+            }
+        }
+        carouselView.completion = completion
+        carouselView.delegate = self
+        self.view = carouselView
+    }
+}
+
+/**
+ Add if you want to track which carousel element has been selected
+ */
+extension EMNotificationViewController: CarouselDelegate {
+    
+    func selectedItem(_ element: EMMessage.Element) {
+        // Add your work...
+        print("Selected element is => \(element)")
+    }
+    
+}