+++ ios/Classes/Constants.swift
@@ -26,6 +26,8 @@ class Constants {
     static let M_SIGNUP = "VL/signUp"
     static let M_GET_EXVISITORID = "VL/getExVisitorID"
     static let M_GET_PUSH_MESSAGES = "RD/getPushMessages"
+    static let M_REQUEST_IDFA = "VL/requestIDFA"
+    static let M_SEND_LOCATION_PERMISSION = "VL/sendLocationPermission"
     
     static let VL_CHANNEL = "IOS"
     static let VL_REQUEST_TIMEOUT = 30