+++ lib/relateddigital_flutter.dart
@@ -24,7 +24,8 @@ class RelateddigitalFlutter {
 
   Future<dynamic> _methodCallHandler(MethodCall methodCall) async {
     if (methodCall.method == Constants.M_TOKEN_RETRIEVED) {
-      RDTokenResponseModel response = RDTokenResponseModel.fromJson(methodCall.arguments);
+      RDTokenResponseModel response =
+          RDTokenResponseModel.fromJson(methodCall.arguments);
       if (_setTokenHandler != null) {
         _setTokenHandler(response);
       }
@@ -35,12 +36,15 @@ class RelateddigitalFlutter {
       }
       _handleUtmParameters(methodCall.arguments);
     } else if (methodCall.method == Constants.M_STORY_ITEM_CLICK) {
-      Map<String, String> map = {'storyLink': methodCall.arguments['storyLink']};
+      Map<String, String> map = {
+        'storyLink': methodCall.arguments['storyLink']
+      };
       _storyPlatformCallbackHandler.onItemClick(map);
     }
   }
 
-  Future<void> init(RDInitRequestModel initRequest, void Function(dynamic result) notificationHandler) async {
+  Future<void> init(RDInitRequestModel initRequest,
+      void Function(dynamic result) notificationHandler) async {
     this.appAlias = initRequest.appAlias;
     this.huaweiAppAlias = initRequest.huaweiAppAlias;
     this._readNotificationHandler = notificationHandler;
@@ -61,9 +65,11 @@ class RelateddigitalFlutter {
     });
   }
 
-  Future<void> requestPermission(Function(RDTokenResponseModel) tokenHandler, {bool isProvisional = false}) async {
+  Future<void> requestPermission(Function(RDTokenResponseModel) tokenHandler,
+      {bool isProvisional = false}) async {
     _setTokenHandler = tokenHandler;
-    await _channel.invokeMethod(Constants.M_PERMISSION, {'isProvisional': isProvisional});
+    await _channel
+        .invokeMethod(Constants.M_PERMISSION, {'isProvisional': isProvisional});
   }
 
   Future<void> setEuroUserId(String userId) async {
@@ -72,28 +78,34 @@ class RelateddigitalFlutter {
 
   Future<void> setEmail(String email, bool permission) async {
     if (email != null && permission != null) {
-      await _channel.invokeMethod(Constants.M_EMAIL_WITH_PERMISSION, {'email': email, 'permission': permission});
+      await _channel.invokeMethod(Constants.M_EMAIL_WITH_PERMISSION,
+          {'email': email, 'permission': permission});
     }
   }
 
   Future<void> setUserProperty(String key, String value) async {
-    await _channel.invokeMethod(Constants.M_USER_PROPERTY, {'key': key, 'value': value});
+    await _channel
+        .invokeMethod(Constants.M_USER_PROPERTY, {'key': key, 'value': value});
   }
 
   Future<void> setAppVersion(String appVersion) async {
-    await _channel.invokeMethod(Constants.M_APP_VERSION, {'appVersion': appVersion});
+    await _channel
+        .invokeMethod(Constants.M_APP_VERSION, {'appVersion': appVersion});
   }
 
   Future<void> setNotificationPermission(bool permission) async {
-    await _channel.invokeMethod(Constants.M_NOTIFICATION_PERMISSION, {'permission': permission});
+    await _channel.invokeMethod(
+        Constants.M_NOTIFICATION_PERMISSION, {'permission': permission});
   }
 
   Future<void> setEmailPermission(bool permission) async {
-    await _channel.invokeMethod(Constants.M_EMAIL_PERMISSION, {'permission': permission});
+    await _channel
+        .invokeMethod(Constants.M_EMAIL_PERMISSION, {'permission': permission});
   }
 
   Future<void> setPhoneNumberPermission(bool permission) async {
-    await _channel.invokeMethod(Constants.M_PHONE_PERMISSION, {'permission': permission});
+    await _channel
+        .invokeMethod(Constants.M_PHONE_PERMISSION, {'permission': permission});
   }
 
   Future<void> setBadgeCount(int count) async {
@@ -104,7 +116,8 @@ class RelateddigitalFlutter {
   }
 
   Future<void> setAdvertisingIdentifier(String advertisingId) async {
-    await _channel.invokeMethod(Constants.M_ADVERTISING, {'advertisingId': advertisingId});
+    await _channel.invokeMethod(
+        Constants.M_ADVERTISING, {'advertisingId': advertisingId});
   }
 
   Future<void> setTwitterId(String twitterId) async {
@@ -112,22 +125,35 @@ class RelateddigitalFlutter {
   }
 
   Future<void> setFacebookId(String facebookId) async {
-    await _channel.invokeMethod(Constants.M_FACEBOOK, {'facebookId': facebookId});
+    await _channel
+        .invokeMethod(Constants.M_FACEBOOK, {'facebookId': facebookId});
   }
 
-  Future<void> customEvent(String pageName, Map<String, String> parameters) async {
-    await _channel.invokeMethod(
-        Constants.M_CUSTOM_EVENT, {'pageName': pageName ?? '', 'parameters': parameters ?? Map<String, String>()});
+  Future<void> customEvent(
+      String pageName, Map<String, String> parameters) async {
+    await _channel.invokeMethod(Constants.M_CUSTOM_EVENT, {
+      'pageName': pageName ?? '',
+      'parameters': parameters ?? Map<String, String>()
+    });
   }
 
-  Future<void> registerEmail(String email, {bool permission = false, bool isCommercial = false}) async {
-    await _channel.invokeMethod(
-        Constants.M_REGISTER_EMAIL, {'email': email, 'permission': permission, 'isCommercial': isCommercial});
+  Future<void> registerEmail(String email,
+      {bool permission = false, bool isCommercial = false}) async {
+    await _channel.invokeMethod(Constants.M_REGISTER_EMAIL, {
+      'email': email,
+      'permission': permission,
+      'isCommercial': isCommercial
+    });
   }
 
-  Future<List> getRecommendations(String zoneId, String productCode, {List filters}) async {
+  Future<List> getRecommendations(String zoneId, String productCode,
+      {List filters}) async {
     String rawResponse = await _channel.invokeMethod(
-        Constants.M_RECOMMENDATIONS, {'zoneId': zoneId, 'productCode': productCode, 'filters': filters ?? []});
+        Constants.M_RECOMMENDATIONS, {
+      'zoneId': zoneId,
+      'productCode': productCode,
+      'filters': filters ?? []
+    });
 
     if (rawResponse != null && rawResponse.isNotEmpty) {
       List result = json.decode(rawResponse);
@@ -149,7 +175,8 @@ class RelateddigitalFlutter {
 
   Future<Map> getFavoriteAttributeActions({String actionId}) async {
     try {
-      String rawResponse = await _channel.invokeMethod(Constants.M_FAV_ATTRIBUTE, {'actionId': actionId});
+      String rawResponse = await _channel
+          .invokeMethod(Constants.M_FAV_ATTRIBUTE, {'actionId': actionId});
 
       if (rawResponse != null && rawResponse.isNotEmpty) {
         Map result = json.decode(rawResponse);
@@ -171,20 +198,23 @@ class RelateddigitalFlutter {
     if (userId == null) {
       return;
     }
-    await _channel.invokeMethod(Constants.M_LOGIN, {'exVisitorId': userId, 'properties': properties ?? Map()});
+    await _channel.invokeMethod(Constants.M_LOGIN,
+        {'exVisitorId': userId, 'properties': properties ?? Map()});
   }
 
   Future<void> signUp(String userId, {Map<String, String> properties}) async {
     if (userId == null) {
       return;
     }
-    await _channel.invokeMethod(Constants.M_SIGNUP, {'exVisitorId': userId, 'properties': properties ?? Map()});
+    await _channel.invokeMethod(Constants.M_SIGNUP,
+        {'exVisitorId': userId, 'properties': properties ?? Map()});
   }
 
   void _handleTokenRegister(RDTokenResponseModel response) {
     Map<String, String> vlTokenParameters = {
       Constants.VL_TOKEN_PARAM: response.deviceToken,
-      Constants.VL_APP_ID_PARAM: response.playServiceEnabled ? appAlias : huaweiAppAlias
+      Constants.VL_APP_ID_PARAM:
+          response.playServiceEnabled ? appAlias : huaweiAppAlias
     };
 
     this.customEvent(Constants.VL_TOKEN_KEY, vlTokenParameters);
@@ -192,12 +222,16 @@ class RelateddigitalFlutter {
 
   void _handleUtmParameters(dynamic payload) {
     try {
-      if (payload != null && payload[Constants.VL_UTM_EVENT_PARAMS_KEY] != null) {
+      if (payload != null &&
+          payload[Constants.VL_UTM_EVENT_PARAMS_KEY] != null) {
         dynamic payloadParams = payload[Constants.VL_UTM_EVENT_PARAMS_KEY];
 
-        String utmCampaign = payloadParams[Constants.VL_UTM_CAMPAIGN_PARAM] as String;
-        String utmSource = payloadParams[Constants.VL_UTM_SOURCE_PARAM] as String;
-        String utmMedium = payloadParams[Constants.VL_UTM_MEDIUM_PARAM] as String;
+        String utmCampaign =
+            payloadParams[Constants.VL_UTM_CAMPAIGN_PARAM] as String;
+        String utmSource =
+            payloadParams[Constants.VL_UTM_SOURCE_PARAM] as String;
+        String utmMedium =
+            payloadParams[Constants.VL_UTM_MEDIUM_PARAM] as String;
 
         if ((utmCampaign != null && utmCampaign.isNotEmpty) ||
             (utmSource != null && utmSource.isNotEmpty) ||
@@ -216,7 +250,9 @@ class RelateddigitalFlutter {
   }
 
   Future<String> getExVisitorID() async {
-    return _channel.invokeMethod<String>(Constants.M_GET_EXVISITORID).then<String>((String value) => value ?? "");
+    return _channel
+        .invokeMethod<String>(Constants.M_GET_EXVISITORID)
+        .then<String>((String value) => value ?? "");
   }
 
   Future<void> sendTheListOfAppsInstalled() async {
@@ -231,16 +267,32 @@ class RelateddigitalFlutter {
 
   Future<PayloadListResponse> getPushMessages() async {
     try {
-      String rawResponse = await _channel.invokeMethod(Constants.M_GET_PUSH_MESSAGES);
+      String rawResponse =
+          await _channel.invokeMethod(Constants.M_GET_PUSH_MESSAGES);
       if (rawResponse != null && rawResponse.isNotEmpty) {
         Map result = json.decode(rawResponse);
         return PayloadListResponse.fromJson(result);
       }
     } on PlatformException catch (e) {
-      return PayloadListResponse([], e.code ?? "" + " : " + e.message ?? "" + " : " + e.stacktrace ?? "");
+      return PayloadListResponse([],
+          e.code ?? "" + " : " + e.message ?? "" + " : " + e.stacktrace ?? "");
     } on Exception catch (ex) {
       print(ex);
       return PayloadListResponse([], ex.toString());
     }
+  }
+
+  Future<void> requestIDFA() async {
+    if (Platform.isAndroid) {
+      if (this._logEnabled) {
+        print('Related Digital - Method not supported on Android');
+      }
+      return;
+    }
+    await _channel.invokeMethod(Constants.M_REQUEST_IDFA);
+  }
+
+  Future<void> sendLocationPermission() async {
+    await _channel.invokeMethod(Constants.M_SEND_LOCATION_PERMISSION);
   }
 }